tests/wpt/tests/css/css-align/parsing/grid-row-gap-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: getComputedStyle().gridRowGap</title>
<link rel="help" href="https://drafts.csswg.org/css-align/#grid-gap-legacy">
<link rel="author" title="Takuya Kurimoto" href="mailto:takuya004869@gmail.com">
<meta name="assert" content="grid-row-gap computed value is a specified keyword or a computed <length-percentage>.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("grid-row-gap", "normal");

test_computed_value("grid-row-gap", "10px");
test_computed_value("grid-row-gap", "20%");
test_computed_value("grid-row-gap", "calc(20% + 10px)");

test_computed_value("grid-row-gap", "calc(-0.5em + 10px)", "0px");
test_computed_value("grid-row-gap", "calc(0.5em + 10px)", "30px");
</script>
</body>
</html>
